pyproject.toml
[project]
name = "fappi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
# TODO: set package versions
dependencies = [
    "alembic",
    "fastapi",
    "email-validator",
    "pydantic",
    "sqlalchemy",
    "bcrypt",
    "python-jose"
]

[dependency-groups]
dev = [
    "fastapi[standard]",
]
test = [
    "pytest",
]

[build-system]
requires = ["uv>=0.6.3"]
